refactor(performances): extract table and form components
