+++ src/components/Performances/PerformanceFormModal.tsx
@@ -1,0 +1,58 @@
+import React from 'react';
+
+import type { CreatePerformanceData } from '@/lib/types/performance';
+import type { Play } from '@/lib/types/play';
+import { FormField, FormModal } from '@/components/ui/FormModal';
+
+export interface PerformanceFormModalProps {
+  isOpen: boolean;
+  isEditing: boolean;
+  plays: Play[];
+  formData: CreatePerformanceData;
+  onChange: (field: string, value: string | number | boolean) => void;
+  onSubmit: (e: React.FormEvent) => void;
+  onClose: () => void;
+  isSubmitting: boolean;
+}
+
+export function PerformanceFormModal({
+  isOpen,
+  isEditing,
+  plays,
+  formData,
+  onChange,
+  onSubmit,
+  onClose,
+  isSubmitting,
+}: PerformanceFormModalProps) {
+  const formFields: FormField[] = [
+    {
+      name: 'play_id',
+      label: 'Obra',
+      type: 'select',
+      required: true,
+      options: plays.map((play) => ({
+        value: play.id,
+        label: play.title || `Obra ${play.id}`,
+      })),
+    },
+    { name: 'date', label: 'Fecha', type: 'date', required: true },
+    { name: 'time', label: 'Hora', type: 'time', required: true },
+    { name: 'location', label: 'Ubicación', type: 'text', required: false },
+  ];
+
+  return (
+    <FormModal
+      isOpen={isOpen}
+      onClose={onClose}
+      title={isEditing ? 'Editar Función' : 'Programar Nueva Función'}
+      icon='🎪'
+      fields={formFields}
+      formData={formData}
+      onChange={onChange}
+      onSubmit={onSubmit}
+      submitText={isEditing ? 'Actualizar Función' : 'Programar Función'}
+      isSubmitting={isSubmitting}
+    />
+  );
+}
